Fixes issue when adding to Sympa ML already subscribed users.
Fixes S3IT issue3567. When adding an email address to the Sympa
mailing list, if this exists already then Sympa returns an error.
While this might be reasonable for a mailing list server, it is not
for us because we *always* try to add the user to the mailing list.

This commit parses the error message and move from "error" to "info"
any message contianing "is already subscribed to the list"...